extract dataflow matcher data structure into a header file
